Made board orientation more general
get_moves figures out what the orientation of the board is from the
grid_to_world function so it knows what direction to take steps in. This
means it doesn't matter what color is on each side.
Also returns the move indexes with respect to top left from baxter's
point of view.