Fixed known bugs in Minimax algorithm so it works
Fixed the issue of the board state getting messed up after a series of
moves were taken. The apparent issue was using self variables in
functions called in the main recursive pruning function. An example was
passing the player's piece number (+/-1) to the make_move and undo_move
functions. The issue that was seen was pieces changing sign mid game.
Fixed it by returning the sign from the get_moves function and passing
it to the make and undo move functions. This seems to have resolved this
issue.

Also fixed indention level of the returns for the recursive function so
it actually prunes the unnecessary branches of the tree, (I think)

Added maximum depth of tree (level)

Needed to reset self variables pertaining to win conditions after a
terminal node was reached so that didn't affect the previous node in the
 tree.