Added basic minimax to test minimax with pruning
When run, both functions return the same move. The algorithm and
updating of alpha and beta is confirmed to be working. However, the
scoring function is weak and needs improvement or else it just returns
the first possible move. Need some randomness and more points/conditions
to get calculate a score from instead of just total piece count.

Current maximum depth of minimax with pruning is 9 layers, any more than
 that and it takes too long to return a move. Max depth without pruning
 is about < 5.