Updated board state scoring for minimax
Added additional scoring points, like relative position of pieces on the
board, where having more pieces in the center is better. Reduced the
storing of the path to the first level, so only the moves from the
initial state are added with their scores to reduce search time later.
Made small change to selection of best move when there are more than one
with the same score.